move snap update logic to base mania DHO
